requirements.txt
speechrecognition
pyttsx3
pyaudio
wikipedia-api
requests
python-dotenv
openai
spacy
psutil
pygame
